# shared/contracts/http/v1/openapi.yaml
openapi: 3.2.0
info:
  title: iHomeland 启动、账号与会话 API
  version: 1.0.0
  summary: 定义 HTTPS 启动、账号、会话与连接 ticket 契约。
servers:
  - url: https://api.example.invalid
    description: 仅用于契约占位；运行 endpoint 必须来自部署配置。
security:
  - bearerAuth: []
paths:
...
  /v1/config:
    get:
      operationId: getBootstrapConfig
      summary: 解析公开限制与连接 endpoint。
      x-ihomeland-body-limit-bytes: 0
      x-ihomeland-timeout-ms: 3000
      x-ihomeland-idempotency: SAFE
      security: []
      responses:
        "200":
          description: 当前公开启动配置。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapConfigResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/auth/register:
    post:
      operationId: registerAccount
      summary: 创建账号与已认证 session。
      x-ihomeland-body-limit-bytes: 4096
      x-ihomeland-timeout-ms: 10000
      x-ihomeland-idempotency: NON_IDEMPOTENT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 账号与 session 已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/auth/login:
    post:
      operationId: loginAccount
      summary: 使用凭据建立新的已认证 session。
      x-ihomeland-body-limit-bytes: 4096
      x-ihomeland-timeout-ms: 10000
      x-ihomeland-idempotency: NON_IDEMPOTENT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 凭据已接受并创建 session。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: 轮换已认证 session 的 token pair。
      x-ihomeland-body-limit-bytes: 4096
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: NON_IDEMPOTENT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token pair 已轮换。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/auth/logout:
    post:
      operationId: logoutSession
      summary: 使当前 session epoch 及其连接失效。
      x-ihomeland-body-limit-bytes: 0
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: IDEMPOTENT
      responses:
        "204":
          description: Session 已失效。
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/session/tickets:
    post:
      operationId: issueConnectionTicket
      summary: 为单一实时 channel 签发一张短期 ticket。
      x-ihomeland-body-limit-bytes: 2048
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: NON_IDEMPOTENT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketRequest"
      responses:
        "201":
          description: 一次性 ticket 已签发。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/world/bootstrap:
    get:
      operationId: getWorldBootstrap
      summary: 查询当前认证 Player 自己的 PersonalWorld 与安全 assignment 投影。
      x-ihomeland-body-limit-bytes: 0
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: SAFE
      responses:
        "200":
          description: 当前 own-world bootstrap 投影。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldBootstrapResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/visits/{visitSessionId}/invites/{inviteId}/accept:
    parameters:
      - $ref: "#/components/parameters/VisitSessionId"
      - $ref: "#/components/parameters/InviteId"
    post:
      operationId: acceptVisitInvite
      summary: 由 bearer 对应的目标 Visitor 接受定向邀请。
      x-ihomeland-body-limit-bytes: 4096
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: IDEMPOTENCY_KEY_REQUIRED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptVisitInviteRequest"
      responses:
        "200":
          description: Invite 已消费并创建或重放同一 reservation。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptVisitInviteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /v1/world/admissions:
    post:
      operationId: issueWorldAdmission
      summary: 为 own-world 或已有 Visitor membership 签发一次性 opaque admission。
      x-ihomeland-body-limit-bytes: 4096
      x-ihomeland-timeout-ms: 5000
      x-ihomeland-idempotency: IDEMPOTENCY_KEY_REQUIRED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorldAdmissionRequest"
      responses:
        "201":
          description: 与当前权威身份、assignment、endpoint 和 purpose 绑定的一次性 admission。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldAdmissionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 在同一认证 actor scope 内稳定标识一次 mutation；相同 key 不得改变目标语义。
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    VisitSessionId:
      name: visitSessionId
      in: path
      required: true
      description: 目标 VisitSession 的安全 ASCII identity，不代表 actor。
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    InviteId:
      name: inviteId
      in: path
      required: true
      description: 目标 invite 的安全 ASCII identity，不包含 credential。
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
  responses:
    ErrorResponse:
      description: 稳定且安全的错误响应。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    BootstrapConfigResponse:
      type: object
      additionalProperties: false
      required: [endpoints, limits]
      properties:
        endpoints:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/Endpoint"
        limits:
          $ref: "#/components/schemas/PublicLimits"
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Za-z0-9](?:[A-Za-z0-9._-]{1,62}[A-Za-z0-9])?$'
          description: 使用 ASCII lowercase canonical key；仅接受 ASCII letter、digit、点、下划线和连字符，且首尾必须为 letter 或 digit。
        password:
          type: string
          minLength: 12
          maxLength: 128
          format: password
          x-ihomeland-max-bytes: 128
          description: 保持原始UTF-8 bytes，不执行trim或Unicode normalization；除字符数约束外最多128 bytes。
        displayName:
          type: string
          minLength: 1
          maxLength: 32
          description: 服务端执行 Unicode NFC、首尾非 control Unicode whitespace 裁剪和连续空白折叠，并拒绝包括 tab、换行在内的 control、format 与 bidi override 字符；长度按规范化后的 Unicode code points 计算。
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Za-z0-9](?:[A-Za-z0-9._-]{1,62}[A-Za-z0-9])?$'
          description: 使用 ASCII lowercase canonical key；仅接受 ASCII letter、digit、点、下划线和连字符，且首尾必须为 letter 或 digit。
        password:
          type: string
          minLength: 1
          maxLength: 128
          format: password
          x-ihomeland-max-bytes: 128
          description: 保持原始UTF-8 bytes，不执行trim或Unicode normalization；除字符数约束外最多128 bytes。
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 2048
    AuthResponse:
      type: object
      additionalProperties: false
      required: [account, session, tokens, endpoints]
      properties:
        account:
          $ref: "#/components/schemas/AccountSummary"
        session:
          $ref: "#/components/schemas/SessionSummary"
        tokens:
          $ref: "#/components/schemas/TokenPair"
        endpoints:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/Endpoint"
    AccountSummary:
      type: object
      additionalProperties: false
      required: [accountId, displayName, createdAtMs]
      properties:
        accountId:
          type: string
          minLength: 1
          maxLength: 64
        displayName:
          type: string
          minLength: 1
          maxLength: 32
        createdAtMs:
          type: integer
          format: int64
          minimum: 0
    SessionSummary:
      type: object
      additionalProperties: false
      required: [sessionId, sessionEpoch, expiresAtMs]
      properties:
        sessionId:
          type: string
          minLength: 1
          maxLength: 128
        sessionEpoch:
          type: integer
          format: int64
          minimum: 1
        expiresAtMs:
          type: integer
          format: int64
          minimum: 0
    TokenPair:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, accessExpiresAtMs, refreshExpiresAtMs]
      properties:
        accessToken:
          type: string
          minLength: 32
          maxLength: 2048
        refreshToken:
          type: string
          minLength: 32
          maxLength: 2048
        accessExpiresAtMs:
          type: integer
          format: int64
          minimum: 0
        refreshExpiresAtMs:
          type: integer
          format: int64
          minimum: 0
    TicketRequest:
      type: object
      additionalProperties: false
      required: [channel]
      properties:
        channel:
          type: string
          enum: [WSS, TLS_TCP]
    TicketResponse:
      type: object
      additionalProperties: false
      required: [ticket, endpoint, scopes, expiresAtMs]
      properties:
        ticket:
          type: string
          minLength: 32
          maxLength: 2048
        endpoint:
          $ref: "#/components/schemas/Endpoint"
        scopes:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            type: string
            enum: [CONTROL, GAMEPLAY]
        expiresAtMs:
          type: integer
          format: int64
          minimum: 0
    PersonalWorldSummary:
      type: object
      additionalProperties: false
      required: [personalWorldId, ownerPlayerId, lifecycle, revision, createdAtMs]
      properties:
        personalWorldId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
        ownerPlayerId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
        lifecycle:
          type: string
          enum: [ACTIVE, ARCHIVED]
        revision:
          type: integer
          format: int64
          minimum: 1
        createdAtMs:
          type: integer
          format: int64
          minimum: 1
    WorldAssignment:
      type: object
      additionalProperties: false
      required: [personalWorldId, worldInstanceId, endpoint, generation, leaseExpiresAtMs]
      properties:
        personalWorldId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
        worldInstanceId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
        endpoint:
          $ref: "#/components/schemas/GameplayEndpoint"
        generation:
          type: integer
          format: int64
          minimum: 1
        leaseExpiresAtMs:
          type: integer
          format: int64
          minimum: 1
    WorldBootstrapResponse:
      type: object
      additionalProperties: false
      required: [world]
      properties:
        world:
          $ref: "#/components/schemas/PersonalWorldSummary"
        assignment:
          $ref: "#/components/schemas/WorldAssignment"
    AcceptVisitInviteRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
...
    AcceptVisitInviteResponse:
      type: object
      additionalProperties: false
      required: [reservation]
      properties:
        reservation:
          $ref: "#/components/schemas/VisitReservation"
    WorldAdmissionRequest:
      oneOf:
        - $ref: "#/components/schemas/OwnWorldAdmissionTarget"
        - $ref: "#/components/schemas/VisitWorldAdmissionTarget"
    OwnWorldAdmissionTarget:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind:
          type: string
          enum: [OWN_WORLD]
...
    WorldAdmissionResponse:
      type: object
      additionalProperties: false
      required: [credential, endpoint, role, purpose, expiresAtMs]
      properties:
        credential:
          type: string
          minLength: 32
          maxLength: 4096
          pattern: '^[A-Za-z0-9._~-]+$'
          description: 仅由 admission verifier 解释的 opaque value，不是公开 claims JSON。
        endpoint:
          $ref: "#/components/schemas/GameplayEndpoint"
        role:
          type: string
          enum: [OWNER, VISITOR]
        purpose:
          type: string
          enum: [OWN_WORLD, JOIN, RECONNECT]
        expiresAtMs:
          type: integer
          format: int64
          minimum: 1
    Endpoint:
      type: object
      additionalProperties: false
      required: [channel, host, port]
      properties:
        channel:
          type: string
          enum: [WSS, TLS_TCP]
        host:
          type: string
          minLength: 1
          maxLength: 253
        port:
          type: integer
          minimum: 1
          maximum: 65535
...
    PublicLimits:
      type: object
      additionalProperties: false
      required: [httpBodyBytes, realtimeFrameBytes]
      properties:
        httpBodyBytes:
          type: integer
          minimum: 1024
          maximum: 1048576
        realtimeFrameBytes:
          type: integer
          minimum: 1024
          maximum: 1048576
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, messageKey, requestId, retryable]
      properties:
        code:
          type: integer
          minimum: 1
        messageKey:
          type: string
          minLength: 1
          maxLength: 128
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        retryable:
          type: boolean
        retryAfterMs:
          type: integer
          minimum: 0
          maximum: 60000
        details:
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/ErrorDetail"